+++ src/commands/newQuartoReprex.ts
@@ -1,8 +1,13 @@
 import * as vscode from "vscode";
-import { QUARTO_WIZARD_LOG } from "../constants";
-import { showLogsCommand } from "../utils/log";
+import { showLogsCommand, logMessage } from "../utils/log";
 import { newQuartoReprex } from "../utils/reprex";
 
+/**
+ * Command to create a new Quarto REPRoducible EXample (reprex).
+ * Prompts the user to select a computing language and then creates a reprex for the selected language.
+ *
+ * @param context - The extension context.
+ */
 export async function newQuartoReprexCommand(context: vscode.ExtensionContext) {
 	const languages = ["R", "Python", "Julia"];
 	const selectedLanguage = await vscode.window.showQuickPick(languages, {
@@ -13,7 +18,7 @@ export async function newQuartoReprexCommand(context: vscode.ExtensionContext) {
 		newQuartoReprex(selectedLanguage, context);
 	} else {
 		const message = `No computing language selected. Aborting.`;
-		QUARTO_WIZARD_LOG.appendLine(message);
+		logMessage(message, "error");
 		vscode.window.showErrorMessage(`${message} ${showLogsCommand()}.`);
 	}
 }