+++ src/utils/extensions.ts
@@ -1,66 +1,18 @@
-import * as vscode from "vscode";
-import * as crypto from "crypto";
 import * as fs from "fs";
 import * as path from "path";
 import * as yaml from "js-yaml";
-import { QUARTO_WIZARD_LOG, QUARTO_WIZARD_EXTENSIONS } from "../constants";
-import { showLogsCommand } from "./log";
+import { logMessage } from "./log";
 
-function generateHashKey(url: string): string {
-	return crypto.createHash("md5").update(url).digest("hex");
-}
-
-export function getGitHubLink(extension: string): string {
-	const [owner, repo] = extension.split("/").slice(0, 2);
-	return `https://github.com/${owner}/${repo}`;
-}
-
-export function formatExtensionLabel(ext: string): string {
-	const parts = ext.split("/");
-	const repo = parts[1];
-	let formattedRepo = repo.replace(/[-_]/g, " ");
-	formattedRepo = formattedRepo.replace(/quarto/gi, "").trim();
-	formattedRepo = formattedRepo
-		.split(" ")
-		.map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-		.join(" ");
-	return formattedRepo;
-}
-
-export async function fetchExtensions(url: string, context: vscode.ExtensionContext): Promise<string[]> {
-	const cacheKey = `${"quarto_wizard_extensions_csv_"}${generateHashKey(url)}`;
-	const cachedData = context.globalState.get<{ data: string[]; timestamp: number }>(cacheKey);
-
-	if (cachedData && Date.now() - cachedData.timestamp < 60 * 60 * 1000) {
-		QUARTO_WIZARD_LOG.appendLine(`Using cached extensions: ${new Date(cachedData.timestamp).toISOString()}`);
-		return cachedData.data;
-	} else {
-		QUARTO_WIZARD_LOG.appendLine(`Fetching extensions: ${url}`);
-	}
-
-	let message = `Error fetching list of extensions from ${QUARTO_WIZARD_EXTENSIONS}.`;
-	try {
-		const response: Response = await fetch(url);
-		if (!response.ok) {
-			message = `${message}. ${response.statusText}`;
-			throw new Error(`Failed to fetch ${url}: ${response.statusText}`);
-		}
-		const data = await response.text();
-		const extensionsList = data.split("\n").filter((line: string) => line.trim() !== "");
-		await context.globalState.update(cacheKey, { data: extensionsList, timestamp: Date.now() });
-		return extensionsList;
-	} catch (error) {
-		QUARTO_WIZARD_LOG.appendLine(`${message} ${error}`);
-		vscode.window.showErrorMessage(`${message}. ${showLogsCommand()}`);
-		return [];
-	}
-}
-
-function findQuartoExtensionsRecurse(dir: string): string[] {
+/**
+ * Recursively finds Quarto extension files in a directory.
+ * @param {string} directory - The directory to search.
+ * @returns {string[]} - An array of file paths to the found extension files.
+ */
+function findQuartoExtensionsRecurse(directory: string): string[] {
 	let results: string[] = [];
-	const list = fs.readdirSync(dir);
+	const list = fs.readdirSync(directory);
 	list.forEach((file) => {
-		const filePath = path.join(dir, file);
+		const filePath = path.join(directory, file);
 		const stat = fs.statSync(filePath);
 		if (stat && stat.isDirectory() && path.basename(filePath) !== "_extensions") {
 			results = results.concat(findQuartoExtensionsRecurse(filePath));
@@ -71,30 +23,46 @@ function findQuartoExtensionsRecurse(dir: string): string[] {
 	return results;
 }
 
-export function findQuartoExtensions(dir: string): string[] {
-	return findQuartoExtensionsRecurse(dir).map((filePath) => path.relative(dir, path.dirname(filePath)));
+/**
+ * Finds Quarto extensions in a directory.
+ * @param {string} directory - The directory to search.
+ * @returns {string[]} - An array of relative paths to the found extensions.
+ */
+export function findQuartoExtensions(directory: string): string[] {
+	return findQuartoExtensionsRecurse(directory).map((filePath) => path.relative(directory, path.dirname(filePath)));
 }
 
-export function getMtimeExtensions(dir: string): { [key: string]: Date } {
-	if (!fs.existsSync(dir)) {
+/**
+ * Gets the modification times of Quarto extensions in a directory.
+ * @param {string} directory - The directory to search.
+ * @returns {{ [key: string]: Date }} - An object mapping extension paths to their modification times.
+ */
+export function getMtimeExtensions(directory: string): Record<string, Date> {
+	if (!fs.existsSync(directory)) {
 		return {};
 	}
-	const extensions = findQuartoExtensions(dir);
-	const extensionsMtimeDict: { [key: string]: Date } = {};
+	const extensions = findQuartoExtensions(directory);
+	const extensionsMtimeDict: Record<string, Date> = {};
 	extensions.forEach((extension) => {
-		extensionsMtimeDict[extension] = fs.statSync(path.join(dir, extension)).mtime;
+		extensionsMtimeDict[extension] = fs.statSync(path.join(directory, extension)).mtime;
 	});
 	return extensionsMtimeDict;
 }
 
-export function findModifiedExtensions(extensions: { [key: string]: Date }, dir: string): string[] {
-	if (!fs.existsSync(dir)) {
+/**
+ * Finds modified Quarto extensions in a directory.
+ * @param {{ [key: string]: Date }} extensions - An object mapping extension paths to their previous modification times.
+ * @param {string} directory - The directory to search.
+ * @returns {string[]} - An array of relative paths to the modified extensions.
+ */
+export function findModifiedExtensions(extensions: Record<string, Date>, directory: string): string[] {
+	if (!fs.existsSync(directory)) {
 		return [];
 	}
 	const modifiedExtensions: string[] = [];
-	const currentExtensions = findQuartoExtensions(dir);
+	const currentExtensions = findQuartoExtensions(directory);
 	currentExtensions.forEach((extension) => {
-		const extensionPath = path.join(dir, extension);
+		const extensionPath = path.join(directory, extension);
 		const extensionMtime = fs.statSync(extensionPath).mtime;
 		if (!extensions[extension] || extensions[extension] < extensionMtime) {
 			modifiedExtensions.push(extension);
@@ -103,6 +71,9 @@ export function findModifiedExtensions(extensions: { [key: string]: Date }, dir:
 	return modifiedExtensions;
 }
 
+/**
+ * Interface representing the data of a Quarto extension.
+ */
 export interface ExtensionData {
 	title?: string;
 	author?: string;
@@ -111,12 +82,17 @@ export interface ExtensionData {
 	source?: string;
 }
 
+/**
+ * Reads a YAML file and returns its data as an ExtensionData object.
+ * @param {string} filePath - The path to the YAML file.
+ * @returns {ExtensionData | null} - The parsed data or null if the file does not exist.
+ */
 function readYamlFile(filePath: string): ExtensionData | null {
 	if (!fs.existsSync(filePath)) {
 		return null;
 	}
 	const fileContent = fs.readFileSync(filePath, "utf8");
-	const data = yaml.load(fileContent) as any;
+	const data = yaml.load(fileContent) as any; // eslint-disable-line @typescript-eslint/no-explicit-any
 	return {
 		title: data.title,
 		author: data.author,
@@ -126,6 +102,12 @@ function readYamlFile(filePath: string): ExtensionData | null {
 	};
 }
 
+/**
+ * Reads Quarto extensions data from a workspace folder.
+ * @param {string} workspaceFolder - The workspace folder to search.
+ * @param {string[]} extensions - An array of extension names to read.
+ * @returns {Record<string, ExtensionData>} - An object mapping extension names to their data.
+ */
 export function readExtensions(workspaceFolder: string, extensions: string[]): Record<string, ExtensionData> {
 	const extensionsData: Record<string, ExtensionData> = {};
 	for (const ext of extensions) {
@@ -139,4 +121,36 @@ export function readExtensions(workspaceFolder: string, extensions: string[]): R
 		}
 	}
 	return extensionsData;
+}
+
+/**
+ * Removes a specified extension and its parent directories if they become empty.
+ *
+ * @param extension - The name of the extension to remove.
+ * @param root - The root directory where the extension is located.
+ * @returns {boolean} - Status (true for success, false for failure).
+ */
+export async function removeExtension(extension: string, root: string): Promise<boolean> {
+	const extensionPath = path.join(root, extension);
+	if (fs.existsSync(extensionPath)) {
+		try {
+			fs.rmSync(extensionPath, { recursive: true, force: true });
+
+			const ownerPath = path.dirname(extensionPath);
+			if (fs.readdirSync(ownerPath).length === 0) {
+				fs.rmdirSync(ownerPath);
+			}
+
+			if (fs.readdirSync(root).length === 0) {
+				fs.rmdirSync(root);
+			}
+			return true;
+		} catch (error) {
+			logMessage(`Failed to remove extension: ${error}`);
+			return false;
+		}
+	} else {
+		logMessage(`Extension path does not exist: ${extensionPath}`);
+		return false;
+	}
 }