+++ src/utils/quarto.ts
@@ -2,11 +2,16 @@ import * as vscode from "vscode";
 import { exec } from "child_process";
 import * as path from "path";
 import * as fs from "fs";
-import { QUARTO_WIZARD_LOG } from "../constants";
-import { findModifiedExtensions, getMtimeExtensions } from "./extensions";
+import { logMessage } from "./log";
+import { findModifiedExtensions, getMtimeExtensions, removeExtension } from "./extensions";
 
 let cachedQuartoPath: string | undefined;
 
+/**
+ * Retrieves the Quarto path from the configuration.
+ *
+ * @returns {string} - The Quarto path.
+ */
 export function getQuartoPath(): string {
 	if (cachedQuartoPath) {
 		return cachedQuartoPath;
@@ -31,6 +36,12 @@ vscode.workspace.onDidChangeConfiguration(async (e) => {
 	}
 });
 
+/**
+ * Checks if the Quarto path is valid and updates the cached path if necessary.
+ *
+ * @param {string | undefined} quartoPath - The Quarto path to check.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the Quarto path is valid, otherwise false.
+ */
 export async function checkQuartoPath(quartoPath: string | undefined): Promise<boolean> {
 	return new Promise((resolve) => {
 		if (!quartoPath) {
@@ -45,6 +56,12 @@ export async function checkQuartoPath(quartoPath: string | undefined): Promise<b
 	});
 }
 
+/**
+ * Checks if the Quarto version is valid by executing the version command.
+ *
+ * @param {string | undefined} quartoPath - The Quarto path to check the version for.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the Quarto version is valid, otherwise false.
+ */
 export async function checkQuartoVersion(quartoPath: string | undefined): Promise<boolean> {
 	return new Promise((resolve) => {
 		exec(`${quartoPath} --version`, (error, stdout, stderr) => {
@@ -57,8 +74,14 @@ export async function checkQuartoVersion(quartoPath: string | undefined): Promis
 	});
 }
 
+/**
+ * Installs a Quarto extension.
+ *
+ * @param {string} extension - The name of the extension to install.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the extension is installed successfully, otherwise false.
+ */
 export async function installQuartoExtension(extension: string): Promise<boolean> {
-	QUARTO_WIZARD_LOG.appendLine(`\n\nInstalling ${extension} ...`);
+	logMessage(`Installing ${extension} ...`, "info");
 	return new Promise((resolve) => {
 		if (vscode.workspace.workspaceFolders === undefined) {
 			return;
@@ -70,7 +93,7 @@ export async function installQuartoExtension(extension: string): Promise<boolean
 
 		exec(command, { cwd: workspaceFolder }, (error, stdout, stderr) => {
 			if (stderr) {
-				QUARTO_WIZARD_LOG.appendLine(`${stderr}`);
+				logMessage(`${stderr}`, "error");
 				const isInstalled = stderr.includes("Extension installation complete");
 				if (isInstalled) {
 					resolve(true);
@@ -89,6 +112,13 @@ export async function installQuartoExtension(extension: string): Promise<boolean
 // This is needed for the extension to be updated in the future
 // To be removed when Quarto supports source records in the _extension.yml file or elsewhere
 // See https://github.com/quarto-dev/quarto-cli/issues/11468
+/**
+ * Updates the _extension.yml file with the source information for the installed extension.
+ *
+ * @param {string} extension - The name of the extension to install.
+ * @param {string} workspaceFolder - The workspace folder path.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the extension source is updated successfully, otherwise false.
+ */
 export async function installQuartoExtensionSource(extension: string, workspaceFolder: string): Promise<boolean> {
 	const extensionsDirectory = path.join(workspaceFolder, "_extensions");
 	const existingExtensions = getMtimeExtensions(extensionsDirectory);
@@ -111,31 +141,47 @@ export async function installQuartoExtensionSource(extension: string, workspaceF
 	return success;
 }
 
+/**
+ * Removes a Quarto extension.
+ *
+ * @param {string} extension - The name of the extension to remove.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the extension is removed successfully, otherwise false.
+ */
 export async function removeQuartoExtension(extension: string): Promise<boolean> {
-	QUARTO_WIZARD_LOG.appendLine(`\n\nRemoving ${extension} ...`);
-
-	return new Promise((resolve) => {
-		if (vscode.workspace.workspaceFolders === undefined) {
-			return;
-		}
-		const workspaceFolder = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-		const quartoPath = getQuartoPath();
-		checkQuartoPath(quartoPath);
-		const command = `${quartoPath} remove ${extension} --no-prompt`;
-
-		exec(command, { cwd: workspaceFolder }, (error, stdout, stderr) => {
-			if (stderr) {
-				QUARTO_WIZARD_LOG.appendLine(`${stderr}`);
-				const isRemoved = stderr.includes("Extension removed");
-				if (isRemoved) {
-					resolve(true);
-				} else {
-					resolve(false);
-					return;
-				}
-			}
-			vscode.commands.executeCommand("quartoWizard.extensionsInstalled.refresh");
-			resolve(true);
-		});
-	});
+	logMessage(`Removing ${extension} ...`, "info");
+	if (vscode.workspace.workspaceFolders === undefined) {
+		return false;
+	}
+	const workspaceFolder = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
+	const status = await removeExtension(extension, path.join(workspaceFolder, "_extensions"));
+	return status;
 }
+/** Quarto command to remove an extension.
+ *
+ * @param extension - The name of the extension to remove.
+ * @param root - The root directory where the extension is located.
+ * @returns {boolean} - Status (true for success, false for failure).
+ *
+ * @deprecated Use removeExtension from extensions.ts instead.
+ */
+// export async function removeExtension(extension: string, root: string): Promise<boolean> {
+// 	return new Promise((resolve) => {
+// 		const quartoPath = getQuartoPath();
+// 		checkQuartoPath(quartoPath);
+// 		const command = `${quartoPath} remove ${extension} --no-prompt`;
+// 		exec(command, { cwd: root }, (error, stdout, stderr) => {
+// 			if (stderr) {
+// 				logMessage(`${stderr}`, "error");
+// 				const isRemoved = stderr.includes("Extension removed");
+// 				if (isRemoved) {
+// 					resolve(true);
+// 				} else {
+// 					resolve(false);
+// 					return;
+// 				}
+// 			}
+// 			vscode.commands.executeCommand("quartoWizard.extensionsInstalled.refresh");
+// 			resolve(true);
+// 		});
+// 	});
+// }