+++ src/utils/reprex.ts
@@ -1,9 +1,14 @@
 import * as vscode from "vscode";
 import * as fs from "fs";
 import * as path from "path";
-import { QUARTO_WIZARD_LOG } from "../constants";
-import { showLogsCommand } from "./log";
+import { showLogsCommand, logMessage } from "./log";
 
+/**
+ * Creates a new Quarto reprex (REPRoducible EXample) file based on the specified language.
+ *
+ * @param {string} language - The programming language for the reprex (e.g., "R", "Julia", "Python").
+ * @param {vscode.ExtensionContext} context - The extension context.
+ */
 export async function newQuartoReprex(language: string, context: vscode.ExtensionContext) {
 	let templateFile = "";
 
@@ -17,18 +22,19 @@ export async function newQuartoReprex(language: string, context: vscode.Extensio
 		case "Python":
 			templateFile = "python.qmd";
 			break;
-		default:
+		default: {
 			const message = `Unsupported language: ${language}.`;
-			QUARTO_WIZARD_LOG.appendLine(message);
+			logMessage(message, "error");
 			vscode.window.showErrorMessage(`${message} ${showLogsCommand()}.`);
 			return;
+		}
 	}
 
 	const filePath = path.join(context.extensionPath, "assets", "templates", templateFile);
 	fs.readFile(filePath, "utf8", (err, data) => {
 		if (err) {
 			const message = `Failed to read the template file: ${err.message}.`;
-			QUARTO_WIZARD_LOG.appendLine(message);
+			logMessage(message, "error");
 			vscode.window.showErrorMessage(`${message} ${showLogsCommand()}.`);
 			return;
 		}